app/requirements.txt
alembic==1.4.3
amqp==5.0.2
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
bcrypt==3.1.7
billiard==3.6.3.0
blinker==1.4
Brlapi==0.7.0
celery==5.0.5
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.4
django-3-jet==1.0.8
django-autoslug==1.9.8
django-celery-results==2.0.1
django-ckeditor==6.0.0
django-crispy-forms==1.7.2
django-jet==1.0.8
django-js-asset==1.2.2
django-model-utils==3.1.2
django-mptt==0.11.0
django-rosetta==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
flake8==3.7.9
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
httplib2==0.14.0
idna==2.10
ipdb==0.13.4
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.4.2
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
keyring==18.0.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
migrate==0.3.8
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
parso==0.8.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
polib==1.1.0
prompt-toolkit==3.0.10
protobuf==3.6.1
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycairo==1.16.2
pycodestyle==2.6.0
pycups==1.9.73
pyflakes==2.1.1
Pygments==2.7.4
PyGObject==3.36.0
PyJWT==2.0.1
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-docx==0.8.10
python-dotenv==0.14.0
python-editor==1.0.4
PythonTurtle==0.3.2
pytz==2020.4
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.34
requests==2.25.1
requests-unixsocket==0.2.0
rosetta==0.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
stories==4.0.0
systemd-python==234
toml==0.10.2
traitlets==5.0.5
turtle==0.0.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
usb-creator==0.3.7
uWSGI==2.0.19.1
vine==5.0.0
virtualenv==20.0.31
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
xkit==0.0.0
